refactor(pricing) Rename Plans component to PricingSection
